RFC-8, Phase Two -- Final submission.  This submission represents the merge of the sandbox/RFC8-2/CsMapDev branch into the trunk.  All modules compile without warnings for all compilers.  All but two modules pass analytic checks per PC-Lint 9.00k without warning.  Exceptions are csIoUtil.cpp (beyond my comprehension) and CS_wellknowntext.cpp (needs total refactoring, thus left to an effort other than a clean-up). This submission includes project files for Visual Studio 2013 (i.e. VC 12.0).
This submission includes some minor tweaks found during final testing.  Note that some more files were deleted, and there are some more name changes.

csBrowser.cpp      -->  Deleted
csmap.rc           -->  Deleted
CS_ed50.c          -->  Deleted
CSmrcomp.c         -->  Deleted

cscscomp.c         -->  CS_cscomp.c
CS_GeoidHeoght.c   -->  CS_geoidHeight.c
CS_VertconUS.c     -->  CS_versonUS.c
CS_COMP.c          -->  CS_Comp.c
csTestCpp.cpp      -->  CS_TestCpp.cpp
csTestCpp.exe      -->  CS_TestCpp.exe


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@2495 4e78687f-474d-0410-85f9-8d5e500ac6b2

+++ CsMapConsoleUtilities/Source/csAscFixer.cpp
@@ -1065,7 +1065,7 @@ int TcsAscDefinition::TrimLineComments (void)
 	TcsDefLnItr lineItr;
 
 	// Remove comment lines from the end of the definition.
-	index = Definition.size () - 1;		// last line
+	index = (int)Definition.size () - 1;		// last line
 	while (index >0)
 	{
 		lineItr = Definition.begin () + index;
